requirements.txt
PyYAML==3.10
Shapely==1.2.18
geonamescache==0.10
kartograph.py==0.6.7
lxml==3.2.3
ordereddict==1.1
pykml==0.1.0
pyproj==1.9.3
pyshp==1.2.0
pytz==2013.7
tinycss==0.3
vincent==0.3.0
